+++ docs/examples/.init.php
@@ -2,25 +2,29 @@
 
 use Oliverde8\Component\PhpEtl\Builder\Factories\ChainSplitFactory;
 use Oliverde8\Component\PhpEtl\Builder\Factories\Extract\CsvExtractFactory;
+use Oliverde8\Component\PhpEtl\Builder\Factories\Extract\JsonExtractFactory;
 use Oliverde8\Component\PhpEtl\Builder\Factories\Grouping\SimpleGroupingFactory;
 use Oliverde8\Component\PhpEtl\Builder\Factories\Loader\CsvFileWriterFactory;
 use Oliverde8\Component\PhpEtl\Builder\Factories\Loader\JsonFileWriterFactory;
 use Oliverde8\Component\PhpEtl\Builder\Factories\Transformer\FilterDataFactory;
 use Oliverde8\Component\PhpEtl\Builder\Factories\Transformer\RuleTransformFactory;
+use Oliverde8\Component\PhpEtl\Builder\Factories\Transformer\SplitItemFactory;
 use Oliverde8\Component\PhpEtl\ChainBuilder;
 use Oliverde8\Component\PhpEtl\ChainOperation\ChainSplitOperation;
 use Oliverde8\Component\PhpEtl\ChainOperation\Extract\CsvExtractOperation;
+use Oliverde8\Component\PhpEtl\ChainOperation\Extract\JsonExtractOperation;
 use Oliverde8\Component\PhpEtl\ChainOperation\Grouping\SimpleGroupingOperation;
 use Oliverde8\Component\PhpEtl\ChainOperation\Loader\FileWriterOperation;
 use Oliverde8\Component\PhpEtl\ChainOperation\Transformer\FilterDataOperation;
 use Oliverde8\Component\PhpEtl\ChainOperation\Transformer\RuleTransformOperation;
+use Oliverde8\Component\PhpEtl\ChainOperation\Transformer\SplitItemOperation;
 use Oliverde8\Component\PhpEtl\ChainProcessor;
 use Oliverde8\Component\PhpEtl\ExecutionContextFactory;
 use Symfony\Component\Yaml\Yaml;
 
 require_once __DIR__ . "/../../vendor/autoload.php";
 
-function getChainProcessor($fileName): ChainProcessor
+function getChainProcessor($fileName, $options = []): ChainProcessor
 {
     $fileName = str_replace(".php", ".yml", $fileName);
 
@@ -43,8 +47,11 @@ function getChainProcessor($fileName): ChainProcessor
     $builder->registerFactory(new CsvFileWriterFactory('csv-write', FileWriterOperation::class));
     $builder->registerFactory(new JsonFileWriterFactory('json-write', FileWriterOperation::class));
     $builder->registerFactory(new CsvExtractFactory('csv-read', CsvExtractOperation::class));
+    $builder->registerFactory(new JsonExtractFactory('json-read', JsonExtractOperation::class));
+    $builder->registerFactory(new SplitItemFactory('split-item', SplitItemOperation::class));
 
     return $builder->buildChainProcessor(
-        Yaml::parse(file_get_contents($fileName))['chain']
+        Yaml::parse(file_get_contents($fileName))['chain'],
+        $options
     );
 }