+++ src/Oliverde8/Component/PhpEtl/ChainBuilder.php
@@ -7,6 +7,7 @@ namespace Oliverde8\Component\PhpEtl;
 use Oliverde8\Component\PhpEtl\Builder\Factories\AbstractFactory;
 use Oliverde8\Component\PhpEtl\ChainOperation\ChainOperationInterface;
 use Oliverde8\Component\PhpEtl\Exception\UnknownOperationException;
+use Symfony\Component\ExpressionLanguage\ExpressionLanguage;
 
 /**
  * Class ChainBuilder
@@ -22,12 +23,16 @@ class ChainBuilder
 
     protected ExecutionContextFactoryInterface $contextFactory;
 
+    protected ExpressionLanguage $expressionLanguage;
+
     /**
      * @param ExecutionContextFactoryInterface $contextFactory
      */
     public function __construct(ExecutionContextFactoryInterface $contextFactory)
     {
         $this->contextFactory = $contextFactory;
+
+        $this->expressionLanguage = new ExpressionLanguage();
     }
 
 
@@ -47,11 +52,11 @@ class ChainBuilder
      * @throws Exception\ChainBuilderValidationException
      * @throws UnknownOperationException
      */
-    public function buildChainProcessor(array $configs): ChainProcessorInterface
+    public function buildChainProcessor(array $configs, array $inputOptions = []): ChainProcessorInterface
     {
         $chainOperations = [];
         foreach ($configs as $id => $operation) {
-            $chainOperations[$id] = $this->getOperationFromConfig($operation);
+            $chainOperations[$id] = $this->getOperationFromConfig($operation, $inputOptions);
         }
 
         return new ChainProcessor($chainOperations, $this->contextFactory);
@@ -63,13 +68,16 @@ class ChainBuilder
      * @throws Exception\ChainBuilderValidationException
      * @throws UnknownOperationException
      */
-    protected function getOperationFromConfig(array $config): ChainOperationInterface
+    protected function getOperationFromConfig(array $config, array $inputOptions): ChainOperationInterface
     {
-        foreach ($this->operationFactories as $factory) {
-            if (is_null($config['options'])) {
-                $config['options'] = [];
+        foreach ($config['options'] as &$option) {
+            if (is_string($option) && strpos($option, "!") === 0) {
+                $option = ltrim($option, '!');
+                $option = $this->expressionLanguage->evaluate($option, $inputOptions);
             }
+        }
 
+        foreach ($this->operationFactories as $factory) {
             if ($factory->supports($config['operation'], $config['options'])) {
                 return $factory->getOperation($config['operation'], $config['options']);
             }