+++ src/Oliverde8/Component/PhpEtl/ChainOperation/Extract/JsonExtractOperation.php
@@ -5,7 +5,6 @@ namespace Oliverde8\Component\PhpEtl\ChainOperation\Extract;
 use oliverde8\AssociativeArraySimplified\AssociativeArray;
 use Oliverde8\Component\PhpEtl\ChainOperation\AbstractChainOperation;
 use Oliverde8\Component\PhpEtl\ChainOperation\DataChainOperationInterface;
-use Oliverde8\Component\PhpEtl\Extract\File\Csv;
 use Oliverde8\Component\PhpEtl\Item\DataItemInterface;
 use Oliverde8\Component\PhpEtl\Item\FileExtractedItem;
 use Oliverde8\Component\PhpEtl\Item\GroupedItem;
@@ -13,27 +12,17 @@ use Oliverde8\Component\PhpEtl\Item\ItemInterface;
 use Oliverde8\Component\PhpEtl\Item\MixItem;
 use Oliverde8\Component\PhpEtl\Model\ExecutionContext;
 
-class CsvExtractOperation extends AbstractChainOperation implements DataChainOperationInterface
+class JsonExtractOperation extends AbstractChainOperation implements DataChainOperationInterface
 {
-    protected string $delimiter;
-
-    protected string $enclosure;
-
-    protected string $escape;
-
     protected string $fileKey;
 
     protected bool $scoped;
 
-    public function __construct(string $delimiter, string $enclosure, string $escape, string $fileKey, bool $scoped)
+    public function __construct(string $fileKey, bool $scoped)
     {
-        $this->delimiter = $delimiter;
-        $this->enclosure = $enclosure;
-        $this->escape = $escape;
         $this->fileKey = $fileKey;
         $this->scoped = $scoped;
     }
-
 
     public function processData(DataItemInterface $item, ExecutionContext $context): ItemInterface
     {
@@ -42,11 +31,12 @@ class CsvExtractOperation extends AbstractChainOperation implements DataChainOpe
             $filename = AssociativeArray::getFromKey($filename, $this->fileKey);
         }
 
-        $fileIterator = new Csv($filename, $this->delimiter, $this->enclosure, $this->escape);
         if ($this->scoped) {
-            $context->getFileSystem()->readStream($filename);
+            $data = json_decode($context->getFileSystem()->read($filename), true);
+        } else {
+            $data = json_decode(file_get_contents($filename), true);
         }
 
-        return new MixItem([new GroupedItem($fileIterator), new FileExtractedItem($filename)]);
+        return new MixItem([new GroupedItem(new \ArrayIterator($data)), new FileExtractedItem($item->getData())]);
     }
 }